requirements.txt
pynput
pyperclip
pyaudio
faster-whisper
torch
numpy
webrtcvad-wheels
pillow
